# docs/swagger.yaml
basePath: /
definitions:
...
  model_configs.CreateModelConfigRequest:
    properties:
      created_by:
        type: string
      is_active:
        type: boolean
      max_tokens:
        type: integer
      mime_type:
        type: string
      model_name:
        type: string
      temperature:
        type: number
      top_k:
        type: number
      top_p:
        type: number
    type: object
  model_configs.ModelConfig:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_tokens:
        type: integer
      mime_type:
        type: string
      model_name:
        type: string
      temperature:
        type: number
      top_k:
        type: number
      top_p:
        type: number
      version:
        type: integer
    type: object
  prompt_templates.CreatePromptTemplateVersionRequest:
    properties:
      created_by:
        type: string
      description:
        type: string
      generation_type:
        type: string
      is_active:
        type: boolean
      metadata:
        type: object
      template:
        type: string
      title:
        type: string
    type: object
  prompt_templates.PromptTemplate:
    description: Prompt template with versioning and activation support
    properties:
      created_at:
        example: "2026-01-19T03:40:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      description:
        example: Classifies documents
        type: string
      generation_type:
        example: CLASSIFICATION
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      metadata:
        type: object
      template:
        example: Classify the following text...
        type: string
      title:
        example: Classification Prompt v1
        type: string
      updated_at:
        example: "2026-01-19T03:40:00Z"
        type: string
      version:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API with password OAuth2 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Learning API
  version: "1.0"
paths:
...
  /model-configs:
    get:
      description: Get all model configurations
      responses:
        "200":
          description: List of model configs
          schema:
            items:
              $ref: '#/definitions/model_configs.ModelConfig'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: List all model configs
      tags:
      - Model Configs
    post:
      description: Create a new model configuration
      parameters:
      - description: Model config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model_configs.CreateModelConfigRequest'
      responses:
        "201":
          description: Created model config
          schema:
            $ref: '#/definitions/model_configs.ModelConfig'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Create model config
      tags:
      - Model Configs
  /model-configs/{id}:
    get:
      description: Retrieve a specific model configuration by its UUID
      parameters:
      - description: Config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Model config details
          schema:
            $ref: '#/definitions/model_configs.ModelConfig'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Config not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get model config by ID
      tags:
      - Model Configs
  /model-configs/{id}/activate:
    post:
      description: Mark a model configuration as active
      parameters:
      - description: Config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Config not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Activate model config
      tags:
      - Model Configs
  /model-configs/active:
    get:
      description: Retrieve the currently active model configuration
      responses:
        "200":
          description: Active model config
          schema:
            $ref: '#/definitions/model_configs.ModelConfig'
        "404":
          description: Active config not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get active model config
      tags:
      - Model Configs
...
  /prompt-templates:
    get:
      description: Get all prompt templates for a specific generation type
      parameters:
      - description: Generation type (CLASSIFICATION or QUESTIONS)
        in: query
        name: generation_type
        required: true
        type: string
      responses:
        "200":
          description: List of prompt templates
          schema:
            items:
              $ref: '#/definitions/prompt_templates.PromptTemplate'
            type: array
        "400":
          description: Bad request - missing generation_type
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: List prompt templates by generation type
      tags:
      - Prompt Templates
    post:
      consumes:
      - application/json
      description: Create a new version of a prompt template (immutable - cannot edit
        existing)
      parameters:
      - description: Template version request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/prompt_templates.CreatePromptTemplateVersionRequest'
      responses:
        "201":
          description: Created prompt template
          schema:
            $ref: '#/definitions/prompt_templates.PromptTemplate'
        "400":
          description: Bad request - invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Create new prompt template version
      tags:
      - Prompt Templates
  /prompt-templates/{id}:
    get:
      description: Retrieve a specific prompt template by its UUID
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Prompt template details
          schema:
            $ref: '#/definitions/prompt_templates.PromptTemplate'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Template not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get prompt template by ID
      tags:
      - Prompt Templates
  /prompt-templates/{id}/activate:
    post:
      description: Mark a prompt template version as active (deactivates other versions)
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Activated prompt template
          schema:
            $ref: '#/definitions/prompt_templates.PromptTemplate'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Activate prompt template
      tags:
      - Prompt Templates
  /prompt-templates/{id}/deactivate:
    post:
      description: Mark a prompt template version as inactive
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Deactivated prompt template
          schema:
            $ref: '#/definitions/prompt_templates.PromptTemplate'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Deactivate prompt template
      tags:
      - Prompt Templates
  /prompt-templates/generation-type/{generationType}:
    get:
      description: Retrieve the currently active prompt template for a specific generation
        type
      parameters:
      - description: Generation type (CLASSIFICATION or QUESTIONS)
        in: path
        name: generationType
        required: true
        type: string
      responses:
        "200":
          description: Active prompt template
          schema:
            $ref: '#/definitions/prompt_templates.PromptTemplate'
        "400":
          description: Bad request - missing generation_type
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Active template not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get active prompt template by generation type
      tags:
      - Prompt Templates
...
securityDefinitions:
  OAuth2:
    flow: password
    tokenUrl: /oauth/token
    type: oauth2
swagger: "2.0"